+++ frontend/src/api/v2/cover.js
@@ -1,24 +1,14 @@
 import { ROOT_PATH } from "./base";
 
-const MISSING_COVER_FN = "missing_cover.webp";
-const MISSING_COVER_PATH = `static/img/${MISSING_COVER_FN}`;
-/*
-export const getCoverSource = (coverPath, updatedAt) => {
-  let src = `${ROOT_PATH}`;
-  src +=
-    coverPath === MISSING_COVER_FN ? MISSING_COVER_PATH : `covers/${coverPath}`;
-  if (updatedAt) {
-    src += `?${updatedAt}`;
-  }
-  return src;
-};
-*/
+const MISSING_COVER_PATH = window.missingCover;
 
 export const getCoverSource = (pk, updatedAt) => {
-  let src = `${ROOT_PATH}`;
+  let src = ROOT_PATH;
   src += pk ? `api/v2/c/${pk}/thumb` : MISSING_COVER_PATH;
-  if (updatedAt) {
+  if (pk && updatedAt) {
     src += `?${updatedAt}`;
+  } else {
+    console.log(src);
   }
   return src;
 };