+++ frontend/src/reader.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import { mapActions, mapGetters } from "vuex";
+import { mapActions, mapGetters, mapMutations } from "vuex";
 
 import ReaderComicPage from "@/components/reader-comic-page";
 import ReaderNavOverlay from "@/components/reader-nav-overlay";
@@ -62,6 +62,7 @@ export default {
         this.routeChanged();
       }
       window.scrollTo(0, 0);
+      this.setBrowseTimestamp();
     },
   },
   created() {
@@ -69,6 +70,7 @@ export default {
   },
   methods: {
     ...mapActions("reader", ["routeTo", "bookChanged", "routeChanged"]),
+    ...mapMutations("browser", ["setBrowseTimestamp"]),
     toggleToolbars: function () {
       this.showToolbars = !this.showToolbars;
     },